Create README.md
This is currently a (somewhat) functional version of my code that I used when writing my MSc dissertation on Quantum Field Theory in Curved Spacetime. This is utilizing Causal Set Theory to calculate the propagator of a massless scalar field on a causal set sprinkled into a spacetime of specified dimension and topology. This code needs a complete refactor so bear with it for now!